preload/preload-edge-cases.html

<!DOCTYPE html>
<title>Test cases for scenarios not defined in the preload spec</title>
<script src="/resources/testharness.js" a></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    function uuidv4() {
        return ([1e7]+-1e3+-4e3+-8e3+-1e11).replace(/[018]/g, c =>
            (c ^ crypto.getRandomValues(new Uint8Array(1))[0] & 15 >> c / 4).toString(16)
        );
    }

    function generateUrl({headers, content, key}) {
        const params = new URLSearchParams(
            {key, content, headers: JSON.stringify(Object.entries(headers || {}))})
        return `resources/inc-requests.py?${params.toString()}`
    }

    async function fetch_loader(url, options) {
        const response = await fetch(url, options)
        return await response.text()
    }

    async function get_count(key) {
        const response = await fetch(`resources/inc-requests.py?key=${key}&count=1`)
        return +(await response.text())
    }

    function preload(url, type = "fetch", crossOrigin) {
        const link = document.createElement('link')
        link.rel = "preload"
        link.href = url
        link.as = type
        if (crossOrigin)
            link.setAttribute("crossorigin", "")
        document.head.appendChild(link)
        return new Promise(resolve => {
            link.addEventListener("load", () => resolve(link))
        })
    }

    const default_load_count = 3
    async function load_and_count({num_loads, content, headers, type, crossOrigin, load, expected}) {
        num_loads = num_loads || default_load_count
        const key = uuidv4()
        const url = generateUrl({headers, key, content})
        const link = await preload(url, type, crossOrigin)
        const results = []
        for (let i = 0; i < num_loads; i++)
            await load(url)

        const count = await get_count(key)
        link.remove()
        assert_equals(count, expected);
    }

    function test_preload_with_fetch({headers, options, num_loads, expected, label}) {
        promise_test(() => 
            load_and_count({headers, num_loads, expected, content: "OK", crossOrigin: true,
                load: url => fetch(url, options)
            }), `Test ${num_loads || default_load_count} fetches after preload with ${label}`)
    }

    function test_preload_with_image({headers, num_loads, expected, label}) {
        promise_test(() => 
            load_and_count({
                headers: {...headers, "Content-Type": "image/svg+xml"},
                num_loads,
                expected, 
                content: '<svg xmlns="http://www.w3.org/2000/svg" width="2" height="2" />',
                type: "image",
                load: url => {
                    const image = document.createElement('img')
                    image.src = url
                    return new Promise(resolve => image.onload = image.onerror = () => {
                        resolve()
                    })
                    document.body.appendChild(image)
                }
            }), `Test ${num_loads || default_load_count} image loads after preload with ${label}`)
    }

    function test_preload_with_script({headers, expected, label, num_loads}) {
        promise_test(() => 
            load_and_count({
                headers: {...headers, "Content-Type": "text/javascript"},
                num_loads,
                crossOrigin: false,
                type: "script",
                content: "//",
                expected, 
                    load: url => {
                        const script = document.createElement('script')
                        script.src = url
                        document.body.appendChild(script)
                        return new Promise(resolve => {
                            script.onload = script.onerror = resolve
                        })

                    }
           }), `Test ${num_loads || default_load_count} script loads after preload with ${label}`)
    }



    const disableCacheResponseHeaders = {"Cache-Control": "no-store"}
    const enableCacheResponseHeaders =  {"Cache-Control": "Max-Age=1000"}

    /*
        Cache disabled

        Chrome+Firefox: preload result is used once
        Safari: preload result is ignored
    */
    test_preload_with_fetch({
        headers: disableCacheResponseHeaders,
        expected: 1,
        label: "cache disabled by server"
    })

    /*
        Cache disabled, but "forced" by client (using force-cache)

        Chrome: preload result is used once
        Firefox+Safari: preload result is ignored
    */
    test_preload_with_fetch({
        headers: disableCacheResponseHeaders,
        options: {cache: "force-cache"},
        expected: 1,
        label: "cache disabled by server, forced by client"
    })

    /*
        Cache enabled using max-age

        Firefox: preload result is used once
        Safari+Chrome: preload result is reused
    */
    test_preload_with_fetch({
        headers: enableCacheResponseHeaders,
        expected: 1,
        label: "cache enabled by server"
    })

    /*
        Cache enabled using max-age + force-cache

        Chrome+Safari+Firefox: preload result is reused
    */
    test_preload_with_fetch({
        headers: enableCacheResponseHeaders,
        options: {cache: "force-cache"},
        expected: 1,
        label: "cache enabled by server + force-cache"
    })

    /*
        Cache enabled image

        Chrome+Safari+Firefox: preload result is reused
    */
    test_preload_with_image({
        headers: enableCacheResponseHeaders,
        expected: 1,
        label: "cache enabled by server"
    })

    /*
        Cache disabled image

        Chrome: preload result is used once
        Safari+Firefox: preload result is reused
    */
    test_preload_with_image({
        headers: disableCacheResponseHeaders,
        expected: 1,
        label: "cache disabled by server"
    })

    /*
        Cache enabled script

        Firefox: preload result is used once
        Chrome+Safari: preload result is reused
    */
    test_preload_with_script({
        headers: enableCacheResponseHeaders,
        expected: 1,
        label: "cache enabled by server"
    })

    /*
        Cache enabled script

        Chrome+Firefox: preload result is used once
        Safari: preload result is reused
    */
    test_preload_with_script({
        headers: disableCacheResponseHeaders,
        expected: 1,
        label: "cache disabled by server"
    })
</script>
<img src="1x1.svg" />
